+++ packages/camera/camera/example/ios/RunnerTests/ThreadSafeMethodChannelTests.m
@@ -7,39 +7,38 @@
 #import <OCMock/OCMock.h>
 
 @interface ThreadSafeMethodChannelTests : XCTestCase
-@property(nonatomic, strong) FlutterMethodChannel *mockMethodChannel;
-@property(nonatomic, strong) FLTThreadSafeMethodChannel *threadSafeMethodChannel;
 @end
 
 @implementation ThreadSafeMethodChannelTests
 
-- (void)setUp {
-  [super setUp];
-  _mockMethodChannel = OCMClassMock([FlutterMethodChannel class]);
-  _threadSafeMethodChannel =
-      [[FLTThreadSafeMethodChannel alloc] initWithMethodChannel:_mockMethodChannel];
-}
-
 - (void)testInvokeMethod_shouldStayOnMainThreadIfCalledFromMainThread {
+  FlutterMethodChannel *mockMethodChannel = OCMClassMock([FlutterMethodChannel class]);
+  FLTThreadSafeMethodChannel *threadSafeMethodChannel =
+      [[FLTThreadSafeMethodChannel alloc] initWithMethodChannel:mockMethodChannel];
+
   XCTestExpectation *mainThreadExpectation = [[XCTestExpectation alloc]
       initWithDescription:@"invokeMethod must be called on the main thread"];
 
-  OCMStub([self.mockMethodChannel invokeMethod:[OCMArg any] arguments:[OCMArg any]])
+  OCMStub([mockMethodChannel invokeMethod:[OCMArg any] arguments:[OCMArg any]])
       .andDo(^(NSInvocation *invocation) {
         if (NSThread.isMainThread) {
           [mainThreadExpectation fulfill];
         }
       });
 
-  [self.threadSafeMethodChannel invokeMethod:@"foo" arguments:nil];
+  [threadSafeMethodChannel invokeMethod:@"foo" arguments:nil];
   [self waitForExpectations:@[ mainThreadExpectation ] timeout:1];
 }
 
 - (void)testInvokeMethod__shouldDispatchToMainThreadIfCalledFromBackgroundThread {
+  FlutterMethodChannel *mockMethodChannel = OCMClassMock([FlutterMethodChannel class]);
+  FLTThreadSafeMethodChannel *threadSafeMethodChannel =
+      [[FLTThreadSafeMethodChannel alloc] initWithMethodChannel:mockMethodChannel];
+
   XCTestExpectation *mainThreadExpectation = [[XCTestExpectation alloc]
       initWithDescription:@"invokeMethod must be called on the main thread"];
 
-  OCMStub([self.mockMethodChannel invokeMethod:[OCMArg any] arguments:[OCMArg any]])
+  OCMStub([mockMethodChannel invokeMethod:[OCMArg any] arguments:[OCMArg any]])
       .andDo(^(NSInvocation *invocation) {
         if (NSThread.isMainThread) {
           [mainThreadExpectation fulfill];
@@ -47,7 +46,7 @@
       });
 
   dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
-    [self.threadSafeMethodChannel invokeMethod:@"foo" arguments:nil];
+    [threadSafeMethodChannel invokeMethod:@"foo" arguments:nil];
   });
   [self waitForExpectations:@[ mainThreadExpectation ] timeout:1];
 }