+++ packages/camera/camera/ios/Classes/FLTThreadSafeEventChannel.h
@@ -8,7 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  * A thread safe wrapper for FlutterEventChannel that can be called from any thread, by dispatching
- * its underlying engine APIs to the main thread.
+ * its underlying engine calls to the main thread.
  */
 @interface FLTThreadSafeEventChannel : NSObject
 
@@ -19,8 +19,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (instancetype)initWithEventChannel:(FlutterEventChannel *)channel;
 
 /*
- * Registers a handler for stream setup requests from the Flutter side on main thread.
- # Completion block runs on main thread.
+ * Registers a handler on the main thread for stream setup requests from the Flutter side.
+ # The completion block runs on the main thread.
  */
 - (void)setStreamHandler:(nullable NSObject<FlutterStreamHandler> *)handler
               completion:(void (^)(void))completion;